refactor: action, reducer 분리
