+++ FE/TAEK/src/store/modules/date/dateReducer.js
@@ -1,8 +1,4 @@
-const CHECK_IN = 'date/CHECK_IN';
-const CHECK_OUT = 'date/CHECK_OUT';
-
-export const checkIn = (date) => ({ type: CHECK_IN, payload: date });
-export const checkOut = (date) => ({ type: CHECK_OUT, payload: date });
+import { CHECK_IN, CHECK_OUT } from './dateAction';
 
 const initialState = {
     checkInDate: null,
@@ -10,7 +6,7 @@ const initialState = {
 }
 
 // react-dates library api 확인 후 보완 할 것 
-export default function date(state = initialState, action) {
+export default function dateReducer(state = initialState, action) {
     switch (action.type) {
         case CHECK_IN:
             return {
@@ -23,8 +19,6 @@ export default function date(state = initialState, action) {
                 checkOutDate: action.payload,
             }
         default:
-            return {
-                ...state
-            }
+            return state
     }
 }
